+++ src/locales/base.ts
@@ -9,6 +9,11 @@ export interface BaseLang {
   edit: string
   reset: string
   close: string
+  back: string
+  clear: string
+  image: {
+    errorTip: string
+  }
   video: {
     errorTip: string
     clickRetry: string
@@ -40,6 +45,8 @@ export interface BaseLang {
     today: string
     loadPreviousMonth: string
     noEarlierMonth: string
+    // eslint-disable-next-line @typescript-eslint/ban-types
+    dayAriaLabel?: Function
   }
   shortpassword: {
     title: string
@@ -142,4 +149,7 @@ export interface BaseLang {
     errorCanvasTips: string
   }
   mask: string
+  quickenter: {
+    title: string
+  }
 }