+++ src/packages/avatar/demos/taro/demo1.tsx
@@ -7,6 +7,8 @@ const Demo1 = () => {
     <Cell className="cell-avatar" align="flex-end">
       <Avatar
         size="60"
+        alt="我是头像"
+        ariaLabel="我是头像"
         src="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
         style={styles}
       />