+++ src/packages/calendarcard/calendarcard.taro.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState, useRef } from 'react'
+import React, { useCallback, useEffect, useState, useRef, useMemo } from 'react'
 import classNames from 'classnames'
 import { View } from '@tarojs/components'
 import { ArrowLeft, ArrowRight, DoubleLeft, DoubleRight } from './icon.taro'
@@ -24,6 +24,7 @@ const defaultProps = {
   ...ComponentDefaults,
   type: 'single',
   firstDayOfWeek: 0,
+  weekdays: [],
 }
 
 const prefixCls = 'nut-calendarcard'
@@ -47,6 +48,8 @@ export const CalendarCard = React.forwardRef<
     renderDay,
     renderDayTop,
     renderDayBottom,
+    renderDayAriaLabel,
+    weekdays,
     onDayClick,
     onPageChange,
     onChange,
@@ -235,6 +238,31 @@ export const CalendarCard = React.forwardRef<
     return d === 0 || d === 6
   }
 
+  const isToday = (day: CalendarCardDay) => {
+    const today = new Date()
+    return (
+      day.year === today.getFullYear() &&
+      day.month === today.getMonth() + 1 &&
+      day.date === today.getDate()
+    )
+  }
+
+  // 日期无障碍朗读
+  const getAriaLabel = (day: CalendarCardDay) => {
+    if (renderDayAriaLabel) {
+      return renderDayAriaLabel(day)
+    }
+    const today = isToday(day)
+    return locale.calendaritem.dayAriaLabel?.(
+      day.year,
+      day.month,
+      day.date,
+      today,
+      isActive(day),
+      isDisable(day)
+    )
+  }
+
   const getClasses = (day: CalendarCardDay) => {
     /**
      * active: single、multiple 激活日期
@@ -397,18 +425,20 @@ export const CalendarCard = React.forwardRef<
     )
   }
 
-  const [weekHeader] = useState(() => {
-    const weekdays = locale.calendaritem.weekdays.map((day, index) => {
+  const weekHeader = useMemo(() => {
+    const weekdaysList =
+      weekdays.length > 0 ? weekdays : locale.calendaritem.weekdays
+    const weekdaysData = weekdaysList.map((day, index) => {
       return {
         name: day,
         key: index,
       }
     })
     return [
-      ...weekdays.slice(firstDayOfWeek, 7),
-      ...weekdays.slice(0, firstDayOfWeek),
+      ...weekdaysData.slice(firstDayOfWeek, 7),
+      ...weekdaysData.slice(0, firstDayOfWeek),
     ]
-  })
+  }, [weekdays, firstDayOfWeek, locale.calendaritem.weekdays])
 
   const renderContent = () => {
     return (
@@ -437,11 +467,14 @@ export const CalendarCard = React.forwardRef<
               )}
               key={`${day.year}-${day.month}-${day.date}`}
               onClick={() => handleDayClick(day)}
+              ariaLabel={getAriaLabel(day)}
+              ariaRole="button"
             >
               <View className={`${prefixCls}-day-top`}>
                 {renderDayTop ? renderDayTop(day) : ''}
               </View>
-              <View className={`${prefixCls}-day-inner`}>
+              {/* @ts-ignore */}
+              <View className={`${prefixCls}-day-inner`} ariaHidden>
                 {renderDay ? renderDay(day) : day.date}
               </View>
               <View className={`${prefixCls}-day-bottom`}>