+++ src/packages/calendarcard/demo.tsx
@@ -12,6 +12,8 @@ import Demo9 from './demos/h5/demo9'
 import Demo10 from './demos/h5/demo10'
 import Demo11 from './demos/h5/demo11'
 import Demo12 from './demos/h5/demo12'
+import Demo13 from './demos/h5/demo13'
+import Demo14 from './demos/h5/demo14'
 
 const CalendarDemo = () => {
   const [translated] = useTranslate({
@@ -30,6 +32,8 @@ const CalendarDemo = () => {
       confirm: '确定',
       ref: '使用 Ref 上的方法',
       title: '搭配 Ref 使用自定义头',
+      customWeekdays: '自定义周几',
+      customAriaLabel: '自定义无障碍朗读',
     },
     'zh-TW': {
       single: '選擇單個日期',
@@ -46,6 +50,8 @@ const CalendarDemo = () => {
       confirm: '確定',
       ref: '使用 Ref 上的方法',
       title: '搭配 Ref 使用自定義头',
+      customWeekdays: '自定義周几',
+      customAriaLabel: '自定义无障碍朗读',
     },
     'en-US': {
       single: 'Select a single date',
@@ -62,6 +68,8 @@ const CalendarDemo = () => {
       confirm: 'Confirm',
       ref: 'Use ref',
       title: 'Custom title',
+      customWeekdays: 'Custom weekdays',
+      customAriaLabel: 'Custom aria label',
     },
   })
 
@@ -92,6 +100,10 @@ const CalendarDemo = () => {
         <Demo11 />
         <h2>{translated.title}</h2>
         <Demo12 />
+        <h2>{translated.customWeekdays}</h2>
+        <Demo13 />
+        <h2>{translated.customAriaLabel}</h2>
+        <Demo14 />
       </div>
     </>
   )