+++ src/packages/popup/popup.tsx
@@ -40,6 +40,8 @@ const defaultProps: WebPopupProps = {
   onTouchStart: () => {},
   onTouchMove: () => {},
   onTouchEnd: () => {},
+  overlayProps: {},
+  animated: true,
 }
 
 // 默认1000，参看variables
@@ -89,6 +91,8 @@ export const Popup: FunctionComponent<
     onTouchMove,
     onTouchEnd,
     closeAriaLabel,
+    overlayProps,
+    animated,
   } = { ...defaultProps, ...props }
   const nodeRef = React.useRef<HTMLDivElement | null>(null)
   const topNodeRef = React.useRef<HTMLDivElement | null>(null)
@@ -106,6 +110,7 @@ export const Popup: FunctionComponent<
   const isTouching = useRef(false)
   const role = 'dialog'
   const { locale } = useConfig()
+  const transitionDuration = animated ? Number(duration) : 0
 
   useLockScroll(nodeRef, innerVisible && lockScroll)
 
@@ -160,7 +165,7 @@ export const Popup: FunctionComponent<
       if (destroyOnClose) {
         setTimeout(() => {
           setShowChildren(false)
-        }, Number(duration))
+        }, transitionDuration)
       }
       onClose && onClose()
     }
@@ -173,7 +178,7 @@ export const Popup: FunctionComponent<
     }
   }
 
-  const handleCloseIconClick = (e: React.MouseEvent<Element, MouseEvent>) => {
+  const handleCloseIconClick = (e: React.MouseEvent | React.KeyboardEvent) => {
     onCloseIconClick(e) && close()
   }
 
@@ -190,9 +195,13 @@ export const Popup: FunctionComponent<
             onClick={handleCloseIconClick}
             role="button"
             aria-label={closeAriaLabel || locale.close}
-            tabIndex={-1}
+            tabIndex={0}
           >
-            {React.isValidElement(closeIcon) ? closeIcon : <Close />}
+            {React.isValidElement(closeIcon) ? (
+              closeIcon
+            ) : (
+              <Close aria-hidden="true" />
+            )}
           </div>
         )}
       </>
@@ -302,7 +311,7 @@ export const Popup: FunctionComponent<
         classNames={transitionName}
         mountOnEnter
         unmountOnExit={destroyOnClose}
-        timeout={duration}
+        timeout={transitionDuration}
         in={innerVisible}
         onEntered={afterShow}
         onExited={afterClose}
@@ -337,8 +346,9 @@ export const Popup: FunctionComponent<
             visible={innerVisible}
             closeOnOverlayClick={closeOnOverlayClick}
             lockScroll={lockScroll}
-            duration={duration}
+            duration={transitionDuration}
             onClick={handleOverlayClick}
+            {...overlayProps}
           />
         )}
         {renderPop()}