+++ src/packages/tabs/demos/taro/demo23.tsx
@@ -9,7 +9,7 @@ const Demo23 = () => {
     {
       title: '自定义 1',
       paneKey: 'c1',
-      icon: <Star />,
+      icon: <Star ariaHidden />,
     },
     {
       title: '自定义 2',