+++ src/types/spec/popup/base.ts
@@ -32,4 +32,6 @@ export interface BasePopup extends BaseProps, BaseOverlay {
   onTouchStart: (height: number, e: any) => void
   onTouchEnd: (height: number, e: any) => void
   closeAriaLabel?: string
+  overlayProps?: any
+  animated?: boolean
 }